split alerticon to its own component
